yay it works

+++ apps/openmw/mwworld/projectilemanager.cpp
@@ -776,7 +776,7 @@ namespace MWWorld
         return false;
     }
 
-    int ProjectileManager::countSavedGameRecords() const
+    size_t ProjectileManager::countSavedGameRecords() const
     {
         return mMagicBolts.size() + mProjectiles.size();
     }
